+++ src/shared/schema-form/services/index.ts
@@ -1,3 +1,4 @@
 export * from "./buildInitialFormState";
 export * from "./executeFieldValidators";
 export * from "./validateSchemaForm";
+export * from "./toFieldEntries";